chore: move styles to their respective variants
