+++ src/components/tabs/Tabs.tsx
@@ -18,91 +18,31 @@ const StyledRoot = styled(Root, {
   variants: {
     theme: {
       light: {
-        'div[role="tab"]': {
-          bg: 'white',
-          '&[data-state="active"]': {
-            color: '$primary',
-            textShadow: '0px 0px 1px currentColor',
-            boxShadow: 'inset 0 -2px 0 0 currentColor'
-          },
-          '&[data-state="inactive"]': {
-            color: '$tonal500'
-          },
-          '&:hover': {
-            color: '$primary',
-            bg: opacify(-0.9, theme.colors.primary.value)
-          },
-          '&:active': {
-            color: '$primary',
-            bg: opacify(-0.8, theme.colors.primary.value),
-            boxShadow: 'none'
-          },
-          '&:focus': {
-            color: '$primary',
-            boxShadow: 'inset 0 0 0 2px currentColor'
-          },
-          '&[data-disabled],&[data-disabled]:hover': {
-            color: '$tonal200',
-            cursor: 'default'
-          }
-        },
-        'div[role="tabpanel"]': {
-          bg: 'white',
-          color: '$textForeground',
-          fontFamily: '$body'
-        },
-        'div[role="tablist"]': {
-          borderBottom: '1px solid $primaryDark'
-        },
-        'button[role="scrollbar"]': {
-          bg: 'white',
-          opacity: 0.8
-        }
+        bg: 'white',
+        color: '$primary'
       },
       dark: {
         bg: '$primaryDark',
-        color: 'white',
-        'div[role="tab"]': {
-          color: 'white',
-          '&[data-state="active"]': {
-            textShadow: '0px 0px 1px currentColor',
-            boxShadow: 'inset 0 -2px 0 0 currentColor'
-          },
-          '&:hover': {
-            bg: opacify(-0.8, 'white')
-          },
-          '&:active': {
-            bg: opacify(-0.7, 'white'),
-            boxShadow: 'none'
-          },
-          '&:focus': {
-            boxShadow: 'inset 0 0 0 2px currentColor'
-          },
-          '&[data-disabled],&[data-disabled]:hover': {
-            color: '$tonal200',
-            cursor: 'default'
-          }
-        },
-        'div[role="tabpanel"]': {
-          bg: '$primaryDark',
-          color: 'white',
-          fontFamily: '$body'
-        },
-        'div[role="tablist"]': {
-          borderBottom: '1px solid white'
-        },
-        'button[role="scrollbar"], button[role="scrollbar"]:focus': {
-          bg: '$primaryDark',
-          opacity: 0.8,
-          color: 'white'
-        }
+        color: 'white'
       }
     }
   }
 })
 
 const StyledTabContent = styled(Content, {
-  flexGrow: 1
+  flexGrow: 1,
+  variants: {
+    theme: {
+      light: {
+        color: '$textForeground',
+        fontFamily: '$body'
+      },
+      dark: {
+        color: 'white',
+        fontFamily: '$body'
+      }
+    }
+  }
 })
 
 export const Tabs: React.FC<TabsProps> & {
